# Optimize lookup_axis1