This optimizes the `lookup_axis1` function and speeds up `differential_rate` by almost a factor 2 for NR data.

When profiling the `differential_rate` function we see that in `lookup_axis1` memory is copied back to the host while the GPU remains idle. The `tf.clip_by_value`, addition and `tf.reshape` is done on the CPU instead of the GPU, this is because the tensor is cast to an integer type (since we later give to `tf.gather`).

By putting the cast after the clip and reshape operations everything is executed on the GPU instead.

Running `differential_rate` on Colab with 1/8 of the NR data and `batch_size=40` the time decreases from 265ms to 140ms.

Also on ER data the differential rate is calculated slightly faster (1.27s from 1.31s) but this is mainly limited by the calculation of the Dirichlet Multinomial.